move table_traits and models into account_activity module
