+++ src/account_activity/mod.rs
@@ -4,3 +4,4 @@ pub mod functions;
 mod graphql;
 pub mod models;
 pub mod page;
+mod table_traits;
