--- src/addresses/mod.rs
@@ -1,2 +1,0 @@
-pub mod models;
-mod table_traits;
